pyproject.toml
[project]
name = "fullmetal-anatomist"
version = "0.1.0"
description = "AI-powered medical study agent that transmutes textbook content into interactive fill-in-the-blank worksheets"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "pydantic>=2.0",
    "pydantic-ai>=0.1.0",
    "python-multipart>=0.0.9",
    "cryptography>=43.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
    "sse-starlette>=2.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
